# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Holivita AI API",
        "description": "API для онбординг-чата с AI-ассистентом, суммаризации и голосового ввода (OpenAI Whisper). Авторизация отключена для MVP — user_id передаётся вручную. Модули: Onboarding (синхронный и асинхронный чат), Summaries (суммаризации сессий), Voice (транскрипция аудио). Формат ответов: успех — {success: true, data: {...}}, ошибка — {success: false, error: \"...\"}.",
        "contact": {
            "name": "Holivita AI Team",
            "email": "support@holivita.ai"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API сервер"
        }
    ],
    "paths": {
        "/onboarding/async/chat": {
            "post": {
                "tags": [
                    "Onboarding Async"
                ],
                "summary": "Асинхронная отправка сообщения в чат онбординга",
                "description": "Ставит сообщение пользователя в очередь для обработки AI-ассистентом. Ответ возвращается немедленно со статусом pending. Для получения результата используйте GET /onboarding/async/status. Если предыдущее сообщение ещё обрабатывается — возвращается ошибка 409. Если message пустое — инициирует начало онбординга.",
                "operationId": "onboardingAsyncChat",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID пользователя",
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 1
                                    },
                                    "session_id": {
                                        "description": "ID сессии (опционально, создаётся автоматически)",
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "message": {
                                        "description": "Текст сообщения. Если пустое — запускается приветствие ассистента.",
                                        "type": "string",
                                        "maxLength": 2000,
                                        "example": "Меня зовут Иван",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Сообщение принято в обработку",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "session_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pending"
                                                    ],
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Предыдущее сообщение ещё обрабатывается",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Предыдущее сообщение ещё обрабатывается."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации входных данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User ID обязателен."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/onboarding/async/status": {
            "get": {
                "tags": [
                    "Onboarding Async"
                ],
                "summary": "Проверка статуса асинхронной обработки сообщения",
                "description": "Возвращает текущий статус обработки последнего сообщения в сессии. Статусы: pending (в очереди), processing (обрабатывается), completed (готово — ответ в поле message), failed (ошибка — описание в поле error). Рекомендуется опрашивать с интервалом 1–2 секунды.",
                "operationId": "onboardingAsyncStatus",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "session_id",
                        "in": "query",
                        "description": "ID сессии онбординга (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус обработки получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "session_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pending",
                                                        "processing",
                                                        "completed",
                                                        "failed"
                                                    ],
                                                    "example": "completed"
                                                },
                                                "message": {
                                                    "description": "Ответ ассистента (только при status=completed)",
                                                    "type": "string",
                                                    "example": "Здравствуйте! Расскажите о себе.",
                                                    "nullable": true
                                                },
                                                "error": {
                                                    "description": "Описание ошибки (только при status=failed)",
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "completed": {
                                                    "description": "true, если обработка завершена успешно",
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Сессия не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Сессия не найдена."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Session ID должен быть валидным UUID."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}